main() function and debug mode created
Main function parses the command line arguments, so that it would be
easy to switch between different modes of calculation. So far, just a
framework for the parameters has been created.

Debug mode involves executing code with pycallgraph. This means, that
pycallgraph.png file will be output with call graph of functions and
their execution profiles.